+++ myems-api/reports/equipmentefficiency.py
@@ -9,7 +9,7 @@ import simplejson as json
 import config
 import excelexporters.equipmentefficiency
 from core import utilities
-from core.useractivity import access_control
+from core.useractivity import access_control, api_key_control
 
 
 class Reporting:
@@ -36,7 +36,12 @@ class Reporting:
     ####################################################################################################################
     @staticmethod
     def on_get(req, resp):
-        access_control(req)
+        if 'API-KEY' not in req.headers or \
+                not isinstance(req.headers['API-KEY'], str) or \
+                len(str.strip(req.headers['API-KEY'])) == 0:
+            access_control(req)
+        else:
+            api_key_control(req)
         print(req.params)
         equipment_id = req.params.get('equipmentid')
         equipment_uuid = req.params.get('equipmentuuid')