+++ myems-api/reports/spaceenergycategory.py
@@ -10,7 +10,7 @@ import gettext
 import config
 import excelexporters.spaceenergycategory
 from core import utilities
-from core.useractivity import access_control
+from core.useractivity import access_control, api_key_control
 
 
 class Reporting:
@@ -41,7 +41,12 @@ class Reporting:
     ####################################################################################################################
     @staticmethod
     def on_get(req, resp):
-        access_control(req)
+        if 'API-KEY' not in req.headers or \
+                not isinstance(req.headers['API-KEY'], str) or \
+                len(str.strip(req.headers['API-KEY'])) == 0:
+            access_control(req)
+        else:
+            api_key_control(req)
         print(req.params)
         space_id = req.params.get('spaceid')
         space_uuid = req.params.get('spaceuuid')