+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added access control to distribution system in myems-api, myems-admin and myems-web
 - added access control to microgrid in myems-api, myems-admin and myems-web
 - added access control to all core objects in myems-api, myems-admin and myems-web
+- added api key access control to all reports in myems-api
 
 ### Changed
 -